+++ analytics/views/base.py
@@ -1,3 +1,5 @@
+import abc
+
 from rest_framework import serializers
 from rest_framework import status
 from rest_framework.exceptions import NotFound
@@ -9,7 +11,7 @@ from rest_framework.views import APIView
 from courses.models import Course
 
 
-class BaseAnalyticsView(APIView):
+class BaseAnalyticsView(APIView, abc.ABC):
     """
     Base view for analytics endpoints.
     Subclasses must define:
@@ -22,26 +24,32 @@ class BaseAnalyticsView(APIView):
     response_serializer_class: type[serializers.Serializer]
 
     def get(self, request: Request) -> Response:
-        course_id = self._validate_request(request)
+        course = self._validate_request(request)
 
-        statistics = self._get_statistics(course_id)
-        response_data = {'course_id': course_id, 'statistics': statistics}
+        statistics = self._get_statistics(course.id)
+        response_data = {
+            'course_public_id': str(course.public_id),
+            'statistics': statistics,
+        }
         return Response(
             self.response_serializer_class(response_data).data,
             status=status.HTTP_200_OK,
         )
 
-    def _validate_request(self, request: Request) -> int:
+    def _validate_request(self, request: Request) -> Course:
         serializer = self.request_serializer_class(data=request.query_params)
         if not serializer.is_valid():
             raise ValidationError(serializer.errors)
 
-        course_id = serializer.validated_data['course_id']
+        course_public_id = serializer.validated_data['course_public_id']
 
-        if not Course.objects.filter(id=course_id).exists():
+        try:
+            course = Course.objects.get(public_id=course_public_id)
+        except Course.DoesNotExist:
             raise NotFound('Course not found.')
 
-        return course_id
+        return course
 
+    @abc.abstractmethod
     def _get_statistics(self, course_id: int) -> list[dict]:
-        raise NotImplementedError
+        pass
